Run futurize -2 -x division_safe -w .

+++ python/lsst/obs/test/testMapper.py
@@ -76,7 +76,7 @@ class TestMapper(CameraMapper):
         @param dataId (dict) Data identifier with visit
         """
         visit = dataId['visit']
-        return long(visit)
+        return int(visit)
 
     def bypass_ccdExposureId(self, datasetType, pythonType, location, dataId):
         return self._computeCcdExposureId(dataId)
